audio: Add support for BT SCO mono streams
Bluetooth SCO's PCM streams are mono but were configured in
stereo to get around a limitation in the McASP driver which
didn't support mono streams.

BT's mono streams require mono-to-stereo and stereo-to-mono
conversions as the primary sound card only supports stereo.
These conversions are in addition to the sample-rate
conversions already in place.

Change-Id: I2712d944f7248b5e4b010620642e79a7378b75d1
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>