audio: primary: Do not set explicit BT PCM thresholds
When the requested period_size and period_count are not
granted by ALSA (i.e. due to hardware rules/constraints),
the pcm_open() call automatically updates those parameters
to the refined values returned by the kernel.

tinyalsa can also update the start_threshold and stop_threshold
accordingly when they are not explicitly set.  The thresholds
set by tinyalsa are correct, except for the stop_threshold
of input streams, which is set to 10x the buffer size.
This value is too large and could hide problems that would
otherwise manifest as overflows (ALSA xruns).

The AudioHAL could also do the params negotiation/update when
the kernel refines those params, but it adds unneeded
complexity that could be avoided by fixing the problem in
tinyalsa.

Change-Id: If690e8c05c2360bfc5eb983427a8adc4d060c645
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>