jacinto6evm: Set permissions for IPC daemon and device files
Add permissions and user/group for:
 * /dev/hwspinlock. It's used to lock/unlock specific hwspinlock
   instances through ioctls
 * /dev/uio0: It's is assumed to be user-space I/O driver for the
   SR0 shared region.
 * /dev/cmem

The user/group is set to media:system as the main users for these
IPC related devices will be audio (which runs as the 'media' user)
and radio (which is assumed to be a system app that runs in the
'system' group).

The LAD daemon user also needs to be changed to 'media' for the
same reasons described above.

Change-Id: Icde29fd022ec83bb0ecc3f4edb185189c4ab3d16
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>