jacinto6evm: configure cpusets for J6 and J6 Eco
By deault only one core (0) is assigned for all categories of cpuset.
But we want both cores (0-1) to be assigned in case of J6.

Since we use same set of binaries for J6 and J6 Eco, configure the
cpusets dynamically using a script.

For J6 - the entries for foreground should be 0-1
For J6 Eco - the entries for foreground should be 0

Change-Id: I1bf8aa5ca4e5b7d73f6fcfdf841f09fa1b4c62df
Signed-off-by: Vishal Mahaveer <vishalm@ti.com>