jacinto6evm: add media_codecs_performance.xml
Some CTS tests require the media_codecs_performance.xml
file to be present on the device in /system/etc in order
for the tests to PASS.

This media_codecs_performance.xml file was generated by
first running the CTS tests on the J6 EVM and then
using the results and the cts-provided get_achievable_rates.py
script.

Note that this file may need to be re-generated for other
devices like J6Eco.

Change-Id: I8c8eaf197493d9c92757d966c46264d227a85eba
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>