audio: Get card name via mixer_get_name()
The card index detection mechanism used in the audio HALs was
based on a customized function which isn't part of the tinyalsa
library.

tinyalsa provides the mixer_get_name() function which also
returns the card name.  The audio HALs now make the card index
detection based on this function and no longer rely on a
customized tinyalsa library.

Change-Id: I10b52df27ad1d4937bfa59a4b2f1b777701c70c0
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>