jacinto6evm: hdmi: Add support for HDMI as primary audio HAL
Add support to build the HDMI audio HAL as primary.  A primary HAL must
support mic capture which the HDMI HAL implements by using the CPU board
microphone input.  The HDMI HAL also supports line-in to be on par with
the CPU board audio HAL.

The primary HAL is still the CPU board based HAL by default, but can be
changed to other supported HALs through the PRIMARY_AUDIO flag in the
device.mk file:
 - 'cpuboard'
 - 'hdmi'
 - 'appe'

Change-Id: I85fc3bed49ac152068ead53881dd00bfda7852b1
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>